redid masking, added new no-mask run, makefile simplified
The bootstrapping is now implemented as part of the mask, which simplifies it and makes it appear in the plots. Because of this I added bootstrapping plots every 10 iterations

There is an option to do the fitting without any masking now too.

A separate plot directory how holds those

The paramter distribution plot is made for the masking and no-masking case